Allow S3 keys to be omitted for instance profiles
Allow the `s3_access_key` and `s3_secret_key` variables to be omitted and
not rendered into the configuration, which allows Docker Registry (and Boto
underneath) to use IAM instance profiles.

This is similar to 799f209 for GCS. However, I'm not providing a separate
variable to enable this, because there are only two options and the naming
(default credentials) wouldn't match the GCS implementation.